Prepare for 019547.SH ytm prediction
