+++ fxincome/const.py
@@ -61,14 +61,83 @@ class DURARION_TYPE(CustomEnum):
     Modified = 'Modified'
 
 
-"""
-Parameters for Mark to Market estimated value
-训练模型的参数。模型目标是预测中债估值收益率涨跌。
-参数主要是Features和Labels
-"""
+class TBOND_PARAM:
+    """
+    Parameters for 16国债19(019547.SH)
+    训练模型的参数。模型目标是预测16国债19收盘收益率的涨跌。
+    参数主要是Features和Labels
+    """
+
+    ALL_FEATS = [
+        # 原始features
+        'date',
+        'close',  # 收盘ytm
+        'amount',  # 成交额，单位是元
+        'ttm',  # Term to Maturity, 剩余期限，单位是年
+        'fr007',  # 7天回购定盘利率
+        't10y',  # 10年国债中债估值ytm
+        'fr007_1y',  # 1年期 fr007 IRS CFETS收盘利率
+        'fr007_5y',  # 5年期 fr007 IRS CFETS收盘利率
+        'usdcny',  # 美元兑人民币汇率
+        # 收盘ytm变种
+        'pct_chg',
+        'avg_chg_5',
+        'avg_chg_20',
+        'volaty',  # (ytm(low price)  - ytm(high price)) / ytm(close)
+        # 流动性指标变种
+        'fr007_chg_5',  # fr007 定盘利率 前5天均值变化率
+        'fr007_1y_chg_5',  # 1年期 fr007 IRS CFETS收盘利率 前5天均值变化率
+        # 收盘ytm与其他各种指标之间的差值
+        'spread_t1y',
+        'spread_t10y',
+        'spread_fr007',
+        'spread_fr007_1y',
+        'spread_fr007_5y',
+        'spread_usdcny',
+        # 其他各种指标之间的差值
+        'spread_fr007_5y_fr007_1y',  # 5年IRS - 1年IRS
+        # 汇率变种
+        'usdcny_chg_5',
+        # label
+        'future',
+        'target'
+    ]
+
+    TRAIN_FEATS = [
+        # 原始features
+        'close',
+        # 'amount',
+        'ttm',
+        # 'fr007',
+        # 'fr007_5y',
+        # 收盘ytm变种
+        'pct_chg',
+        # 'avg_chg_5',
+        'avg_chg_20',
+        # 'volaty',
+        # 流动性指标变种
+        'fr007_chg_5',
+        # 收盘ytm与其他各种指标之间的差值
+        # 'spread_t1y',
+        'spread_t10y',
+        # 'spread_fr007',
+        # 'spread_fr007_5y',
+        # 'spread_usdcny',
+        # 其他各种指标之间的差值
+        # 'spread_fr007_5y_fr007_1y',
+        # 汇率变种
+        'usdcny_chg_5'
+    ]
+    LABELS = ['target']
 
 
 class MTM_PARAM:
+    """
+    Parameters for Mark to Market estimated value
+    训练模型的参数。模型目标是预测中债估值收益率涨跌。
+    参数主要是Features和Labels
+    """
+
     ALL_FEATS = [
         # 原始features
         'date',
